Issue #310: Issue with maxYear
"When I set maxYear to new Date().getFullYear() (now is 2018)
in options then in UI datepicker I can select only until 2017."
The same does not happen to minYear.
So I updated the behavior of maxYear to match minYear's.

"Also, when I swap months one by one it can swap any year.
Why it does not apply max Year?"
There is a limitation on nextMonth and prevMonth functions as they
are not checking for the lower and upper limits (when defined).